# go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Who knows",
        "version": "3.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "description": "Authenticates a user using username and password",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful registration",
                        "schema": {
                            "$ref": "#/definitions/models.AuthResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPValidationError"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "get": {
                "description": "Displays the login page",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Show login page",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AuthResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "models.HTTPValidationError": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ValidationErrorDetail"
                    }
                }
            }
        },
...
        "models.ValidationErrorDetail": {
            "type": "object",
            "properties": {
                "loc": {
                    "type": "array",
                    "items": {}
                },
                "msg": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    }
}